requirements.txt
requests
BeautifulSoup4==4.10.0
Pillow==9.0.1